fix mismatched types int64 and uint64

+++ br/pkg/lightning/common/storage_unix.go
@@ -51,7 +51,7 @@ func GetStorageSize(dir string) (size StorageSize, err error) {
 	}
 
 	// Available blocks * size per block = available space in bytes
-	size.Available = stat.Bavail * bSize
+	size.Available = uint64(stat.Bavail * bSize)
 	size.Capacity = stat.Blocks * bSize
 
 	return
